apka/requirements_TO.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==24.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.9
docopt==0.6.2
einops==0.8.0
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
huggingface-hub==0.27.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.5.0
MouseInfo==0.1.3
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
opencv-python==4.10.0.84
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
proglog==0.1.10
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pyAudioAnalysis==0.3.14
PyAutoGUI==0.9.54
pycparser==2.22
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.0
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.1
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.1.10
yarl==1.18.3
